+++ enclone_main/src/lib.rs
@@ -7,5 +7,6 @@ pub mod flag_defective;
 pub mod inconsistent;
 pub mod main_enclone;
 pub mod merge_onesies;
+pub mod populate_features;
 pub mod setup;
 pub mod split_orbits;
